Add permission object and permission lookups for documents.
(This allows documents to, in theory, have their own permissions
separately from the folders they are in.  It also allows for the same
permissions on a folder mapped to a (to-be-created) Owner role to affect
different documents in the folder in different ways.  But for now none
of that is exposed by the UI.)


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@3510 c91229c3-7414-0410-bfa2-8a42b809f60b