Pass the redirect around on errors.
Do the cookie test in a separate action, which is redirected to after
login.  This has the advantage of not causing problems when you open on
login window, then another, and then log in with the first login window
(since the cookie set by the second login window overwrote it).


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@5212 c91229c3-7414-0410-bfa2-8a42b809f60b