Merged in from DEV trunk...
"KTS-1377"
"Not to delete roles when overriding inherited assigning of roles"
Fixed. When the role is created, the users and groups should be inherited from the parent role.

Reviewed By: Kevin Fourie

"KTS-2023"
"Fatal Error when clicking 'backup now' btn on the Upgrade Wizard"
Fixed.

Reviewed By: Kevin Fourie

"KTS-1303"
"Plugin and indexer paths are saved in the DB"
Fixed.

Reviewed by: Kevin Fourie

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@6712 c91229c3-7414-0410-bfa2-8a42b809f60b