Merged in from DEV trunk...
"KTS-1526"
"A Required field not being marked as such leads to a DB error: 'DB Error: null value violates not-null constraint'"
Fixed. Made the workflowid part of the form post.

Reviewed By: Kevin Fourie

"KTS-1526"
"A Required field not being marked as such leads to a DB error: 'DB Error: null value violates not-null constraint'"
Fixed. Added a hidden widget for forms.

Reviewed By: Kevin Fourie

"KTS-1526"
"A Required field not being marked as such leads to a DB error: 'DB Error: null value violates not-null constraint'"
Fixed. Added a hidden widget for forms.

Reviewed By: Kevin Fourie

"KTS-1526"
"A Required field not being marked as such leads to a DB error: 'DB Error: null value violates not-null constraint'"
Fixed. Added a hidden widget for forms.

Reviewed By: Kevin Fourie

"KTS-1526"
"A Required field not being marked as such leads to a DB error: 'DB Error: null value violates not-null constraint'"
Fixed. Added a hidden widget for forms and created the key for the workflow reference on the first page.

Reviewed By: Kevin Fourie

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@6725 c91229c3-7414-0410-bfa2-8a42b809f60b