Merged in from DEV trunk...
KTS-2591
"When updating folder permissions for a folder with an & in its name the progresstext appears as '&amp'"
Fixed. The message is being htmlentitied for cross-scripting, added a str_replace for the &amp;.

Reviewed by: Conrad Vermeulen
Committed by: Megan Watson

KTS-2599
"Indexing files with no extractors associated results in the index queue filling up"
Fixed. 

Committed By: Conrad Vermeulen
Reviewed By: Kevin Fourie

KTS-2584
"Postscrip extractor is disabled on linux"
Fixed. The issue was that the full path to the application was not used and that the parameters were incorrect.

Committed By: Conrad Vermeulen
Reviewed By: Megan Watson

BBS-1059 / SUP-411
"On the 3.4.2 Version of KTTools, reason is mandatory field for each action, i would like to desactivate this."
Added new client policies for reasons.

Committed By: Isaac Lundall
Reviewed By: Conrad Vermeulen

KTS-2618
"#jsonlookup.js# is blocking migration to launchpad and is a temporary file that is not required"
Deleted.

Committed By: Conrad Vermeulen
Reviewed By: Daniel Chalef


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@7592 c91229c3-7414-0410-bfa2-8a42b809f60b