Merged in from STABLE trunk...
KTS-2198
"When setting document permissions on a workflow state the workflow has to be reset for all files in repository for this to take affect"

Fixed
Permissions should now be reset on all documents associated with that state automatically.

Committed By: Jonathan Byrne
Reviewed By: Jalaloedien Abrahams

KTS-2630
"Users with only READ, WRITE and FOLDER_DETAILS permission can create folder with the BULK_UPLOAD functionality."

Fixed
Added a permission check that won't allow a folder to be created if the user does not have permission to create one.

Committed By: Jonathan Byrne
Reviewed By: Jalaloedien Abrahams

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@7629 c91229c3-7414-0410-bfa2-8a42b809f60b