Merged in from DEV trunk...
KTS-2912
"Document Indexing Diagnostics shows that there is a problem indixing a PDF file"
Fixed.

Commited By: Conrad Vermeulen
Reviewed By: Jonathan Byrne

KTS-2892
"Search for user by username returns all users and filter is ignored."
Fixed. The sql function to get the list of users was overwriting the filter.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen

KTS-2873
"Double quote (") is not accepted as valid input for metadata when the name of a Lookup field or a Tree field has a Double quote in."
Fixed. Sanitised the values.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen

KTC-331
"Bulk upload throws an exception shown on attached screenshot"
Fixed. The upload is successful regardless of the exception, put @ in front of the chgrp() to suppress the error message.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen



git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@8031 c91229c3-7414-0410-bfa2-8a42b809f60b