Merged in from DEV trunk...
KTS-3043
"On reread of plugins the quicklinks plugin is set to disabled"
Fixed. Moved the checks on the end version after an upgrade into the if statement containing the upgrade.

Committed By: Megan Watson
Reviewed By: Conrad Vermeulen

KTS-2913
"Search2: A search query for a FOLDER returns no results"
Fixed. Also ensured that wildcards are catered for correctly for LIKE operations.

Committed By: Conrad Vermeulen
Reviewed By: Kevin Fourie

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@8078 c91229c3-7414-0410-bfa2-8a42b809f60b