Merged in from DEV trunk...
KTC-399
"Bulk checkout: when you checkout an already checked out document you get an information alert that is not clear"
Fixed. If a document is checked out by the current user it now gets added to the downloaded files.

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-2655
"Error Message On Removing Permissions but action goes ahead anyway"
Fixed. Added a check to ensure the user doesn't removed his/her own permission to manage security.

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-2873
"Double quote (") is not accepted as valid input for metadata when the name of a Lookup field or a Tree field has a Double quote in."
Fixed. The values in the sub categories needed to be htmlentities().

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-3091
"The name of the open office service needs to be changed in the script to check if its running"
Fixed. Changed the name and the paths to match the new stack.

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-3093
"Creating fieldsets and fields with quotes results in double escaping of names and descriptions"
Fixed.

Committed By: Conrad Vermeulen
Reviewed By: Megan Watson

KTS-3096
"Creating workflow, workflow states, workflow transition names with quotes results in double escaping of names and descriptions"
Fixed.

Committed By: Conrad Vermeulen
Reviewed By: Megan Watson

KTS-3097
"Field names with quotes was impacting on search2 query builder"
Fixed.

Committed By: Conrad Vermeulen
Reviewed By: Megan Watson

KTS-3085
"Storage utilisation plugin does not seem to be working on Windows Vista."
Fixed. Unfortunately, df does not work on Vista. We now detect if df has a problem and disable the dashlet accordingly.

Committed By: Conrad Vermeulen
Reviewed By: Megan Watson

KTC-403
"One is able to make a checked out document immutable"

Fixed. Added a check to return an error if the document is checked out.

Committed By: Jonathan Byrne
Reviewed By: Megan Watson

KTS-1888
"Expunge page in the DMS Administration section, pagination not working, links to consecutive pages do not work"

Fixed. Changed the php file to return the contents of a div and not just a table body so that pagination works in Internet Explorer.

Committed By: Jonathan Byrne
Reviewed By: Jalaloedien Abrahams


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@8171 c91229c3-7414-0410-bfa2-8a42b809f60b