Merged in from DEV trunk...
KTC-395
"Unable to compare content versions for a txt doc or an Open Office document"
Fixed. Changed the way it checks if open office is running.

Committed By: Megan Watson
Reviewed By: Conrad Vermeulen

KTS-3104
"Dashboard Info Dashlets displaying white blocks when rolled up."

Fixed. Added new pngs to be used when an info dashlet is rolled up and changed css accordingly.

Committed By: Jonathan Byrne
Reviewed By: Jalaloedien Abrahams

KTS-2885
"Positioning of where the notification icon appears on dashlets is incorrect. Currently vertically aligned middle and should be text-top."

Fixed. Changed the positioning of the error and info dashlet notifictation icons to top.

Committed By: Jonathan Byrne
Reviewed By: Jalaloedien Abrahams

KTS-3106
"On the log-in/logout dialog and the footer of the KTDMS web browser interface there is no version description added to the version."
Fixed. If the wintools plugin is not enabled, Community Edition is appended to the version name.

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-2695
"The character Ñ is replaced by ? when used in a document filename in the DMS"

Fixed. Changed smarty wordwrap to make a new line on the end of a word to stop question marks from displaying.

Committed By: Jonathan Byrne
Reviewed By: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@8194 c91229c3-7414-0410-bfa2-8a42b809f60b