Merged in from DEV trunk...
KTS-3161
"Allow the user to change the ordering of the browse columns"
Fixed. Added the logic to change the order.

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-3162
"Allow the user to change the ordering of the fields in a fieldset"
Fixed. Added columns and logic for ordering the fields

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-3231
"No email sent when you are subscribed to a folder or a document."
Fixed. Removed the text on the icon.

Committed by: Megan Watson
Reviewed by: Jonathan Byrne

KTS-3234
"CLONE -When a document starts of as e.g. a .doc file, but the next version is a pdf file, then it remains to be seen as a Word document in some respects(SUP-675)"
Fixed. Added a mimetype id change if the filename changes in the rename function.

Committed By: Jonathan Byrne
Reviewed By: Megan Watson


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/trunk@8354 c91229c3-7414-0410-bfa2-8a42b809f60b