Merged in from DEV trunk...
KTS-3307
"The Config file data needs to be moved to the database and an admin page needs to be made to manage this data."
Fixed. Made changes to run config from DB.

KTS-3309
"Dmsdefaults has a lot of config functions in it, they need to be moved to the KTConfig class."
Fixed. Functions have been separated into the right classes.

KTS-3272
"Migration of configuration to database"
Fixed. Completed KTS-3276 and KTS-3278 associated with this issue.

Committed By: Jonathan Byrne
Reviewed By: Megan Watson

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/branches/3.5.3-Branch@8565 c91229c3-7414-0410-bfa2-8a42b809f60b