Merged in from STABLE trunk..
KTS-3234
"CLONE -When a document starts of as e.g. a .doc file, but the next version is a pdf file, then it remains to be seen as a Word document in some respects(SUP-675)"
Fixed. The mime type is only changed if the filename changes and is derived from the new filename.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen

Plus various updates that did not make the last commit.

git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.3a-Release-Branch@9259 c91229c3-7414-0410-bfa2-8a42b809f60b