Merged in from STABLE trunk...
KTS-3688
"Refactor kt_url() and review the cache/serverName.txt implementation"
Fixed. The serverName is now stored in the DB and only used when $_SERVER['HTTP_HOST'] is empty or set to localhost (the default in dmsDefaults). The dynamicConfigSettings are resolved every time and not cached to prevent localhost being cached as the serverName.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.3a-Release-Branch@9332 c91229c3-7414-0410-bfa2-8a42b809f60b