Merged in from STABLE trunk...
KTS-3721
"When attempting to access the tag from the Dashboard Tag Cloud dashlet"
Fixed. The url needed to be encoded.

KTS-3720
"When using the following characters as a tag 'Â´`&;!.~$()#=[]^@-_{}Â°ÂºÂ©ÂºÂªâªÂ¿âÂ¬Â½Â¼Â¾Â±Â¡âÂ£Â¥â§ Â«Â» the document does not show when the tag is navigated and the system incorrectly reports that 2 additional tags were created for the document."
Fixed. The tag in redirect url is now encoded and decoded.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.3a-Release-Branch@9361 c91229c3-7414-0410-bfa2-8a42b809f60b