Merged in from STABLE trunk...
KTC-582
"UI defect on I.E 7 Windows XP and RedHat Linux"
Fixed. Replaced the <p> tags with div's which render the css correctly. Set the permissions table width to 60% to allow the list to adjust its width properly.

KTS-3344
"Make provision for very long Firstname and Lastname combinations in the caption bar for KTDMS web interface."
Fixed. Set the truncate to 40 characters to account for the admin menu on a 1024x768 res.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.3a-Release-Branch@9376 c91229c3-7414-0410-bfa2-8a42b809f60b