Merged in from STABLE trunk...
KTS-3716
"Document size indication appears incorrect on the 'Browse' tab for a document which has the Right-to-Left characters as its name."
Fixed. Replaced the spans with divs which separates the link and the right-to-left characters from the mime icon and size so they render within the div and don't affect the mime icon and size.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.3a-Release-Branch@9378 c91229c3-7414-0410-bfa2-8a42b809f60b