Merged in from STABLE trunk...
KTC-594
"When a shortcut is made to an XML, RTF, TXT or CSS file there is no indication that it is a shortcut after the process has completed."
Fixed. Added the folder mime icon and the generic icon. The shortcut css class are now .folder_shortcut not .folder.shortcut - IE gets confused sometimes.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.3a-Release-Branch@9392 c91229c3-7414-0410-bfa2-8a42b809f60b