KTS-3691 "CLONE -Is there a way to prevent from checking in the wrong file?(SUP-1064)" Fixed. Added a config option to hide the checkbox for "Force Original Filename".
KTS-3736
"User should be made aware that setting the field limit for document filenames and titles beyond 255 is futile as only 255 characters are allowed in the fields in the DB."
Fixed. Added a better description of the config option.

KTC-592
"Better default tooltip needed on mouse over than 'ACME Corporation'."
Fixed. Added a better tooltip as the default value for the config option.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen



git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@9396 c91229c3-7414-0410-bfa2-8a42b809f60b