Merged in...
KTS-3738
"Option for 'PDF Generator' available in Edit mode for the 'Action Restrictions' of a Workflow but cannot be seen when the edit of the 'Action Restrictions' is saved"
Fixed. The edit was using getName() if displayName() returned empty. Fixed displayName() to return the name if the document object is null but python is installed.

KTS-3737
"Although the Action Restrictions of a Workflow state prohibit 'PDF Generator' the 'Generate PDF' action can still be seen in the Document Actions of the document."
Fixed. Removed the getName() function that was overriding the default function which returns the namespace.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.4-Release-Branch@9437 c91229c3-7414-0410-bfa2-8a42b809f60b