Merged in...
Updated i18n files.

KTS-3770
"Error on custom error page show incorrect message because it is not removed from the session"
Implemented.

KTS-3755
"When navigating the folder after it has been found in a seach (Advanced Search) navigates the user to the Root folder and not the folder itself."
Fixed.

Committed By: Conrad Vermeulen
Reviewed By: Megan Watson

KTS-3741
"The Property Preview Panel lacks a scroll bar"
Fixed. Moved the overflow to the panel window instead of the content div.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen


git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.4-Release-Branch@9445 c91229c3-7414-0410-bfa2-8a42b809f60b