KTS-3774 "Although 'Action Restrictions' for 'Archive' is set the user can still 'Archive' the document via the bulk actions when in folder view." Fixed. Added a check for workflow before doing the archive.
KTS-3758
"When attempting to Archive a folder that contains no documents the Status is given as being 'Non-numeric identifier'"
Fixed. Added a check around explode($sDocuments) to create an empty array if $sDocuments is empty otherwise it creates an array containing an empty array which is not seen as empty.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen



git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@9468 c91229c3-7414-0410-bfa2-8a42b809f60b