Merged in...
KTC-621
"Deleting a folder causes the error: PEAR_Error::getSymbolicLinks()"
Fixed. Moved the loop to get the symbolic links above the code to delete the folder.

KTS-3774
"Although 'Action Restrictions' for 'Archive' is set the user can still 'Archive' the document via the bulk actions when in folder view."
Fixed. Added a check for workflow before doing the archive.

KTS-3758
"When attempting to Archive a folder that contains no documents the Status is given as being 'Non-numeric identifier'"
Fixed. Added a check around explode($sDocuments) to create an empty array if $sDocuments is empty otherwise it creates an array containing an empty array which is not seen as empty.

KTS-3780
"When cancelling / returning from a bulk action in the tag cloud search the user is taken to the browse page"
Fixed. Changed the return action to contain the full link. Added a default in getReturnUrl() in bulk action to use the full url.

KTS-3757
"After a Bulk action was performed on folders in the Search results pane/view the user is returned to the original view where the folders search is shown in its previous state."
Fixed. On completing the bulk action the return url calls refresh which repeats the search.

KTS-3759
"When the user does not select a folder or documenton the search result view and uses any of the Bulk Actions he is returned to the Root fodler view."
Fixed. It now builds the return url based on the request parameters.

KTS-3764
"Items in admin section registered more than once."
Fixed. Added a lock around the plugin registration (reread) to prevent a race condition.

Committed by: Megan Watson
Reviewed by: Conrad Vermeulen



git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/STABLE/branches/3.5.4-Release-Branch@9476 c91229c3-7414-0410-bfa2-8a42b809f60b