Create kt.api.preload function to deal with all fragment and exec preloading.
Inside this function use a regular for loop, instead of 'for ... in' which
appeared to parse an additional element not added by the code.  Regular for
loop only gets the elements we explicitly add and no phantom elements.

We can now replace all 'for...in' fragment/exec loading loops with a call to
kt.api.preload.

Changed loading in invite and share js to use the new preload function.

Story/Branch/Issue: webcomms_refactor

Committed by: Paul Barrett