Invite user window now correctly caches preloaded content.
Other fragment caching is still not working (at least not shared users and
upload dialog.)

This brings up the question: do we want to preload and cache all this stuff?
Some of it might not be relevant to preloading.

Still going to try to get the rest to cache correctly IF preloaded, anyway.

Ref: webcomms_refactor

Committed by: Paul Barrett