Priority Queues
Added an option to ktapi's add_document function to indicate ktapi as the source of the add.
Added the options array to the indexer::index function to determine the source of the add.
Created a util map function to set priorities for different process types.
Added priorities for indexing, processing and email.

Committed by: Megan Watson