Modal document / bulk actions Incl. copy/move story
Added the document names into the confirmation actions.
Changed the confirmation dialog to display the correct action.
Added inputs for changing the title / filename on a move with a conflict.
Added js function to upper case the first letter of a word.
Changed the urls to clean urls.

Committed by: Megan Watson