Add script for easy branch deletion.
Given a branch name, the script deletes the remote and any local copies.

Does not check whether the local copy exists, so you will get an error if it does not.

Uses the -D switch to force deletion.  Be sure you want to delete the local copy.

Ref: Git scripts

Committed by: Paul Barrett